move admin validation survey actions to separate file
remove copyright in the middle of a file + run yarn format
