+++ packages/evolution-legacy/src/components/admin/ValidateSurvey.js
@@ -16,7 +16,7 @@ import Section                from '../survey/Section';
 import SectionNav             from '../survey/SectionNav';
 import { withSurveyContext } from 'evolution-frontend/lib/components/hoc/WithSurveyContextHoc';
 import LoadingPage from 'chaire-lib-frontend/lib/components/pages/LoadingPage';
-import { startSetSurveyValidateInterview, startUpdateSurveyValidateInterview, startSurveyValidateAddGroupedObjects, startSurveyValidateRemoveGroupedObjects } from 'evolution-frontend/lib/actions/Survey';
+import { startSetSurveyValidateInterview, startUpdateSurveyValidateInterview, startSurveyValidateAddGroupedObjects, startSurveyValidateRemoveGroupedObjects } from 'evolution-frontend/lib/actions/SurveyAdmin';
 import { InterviewContext } from 'evolution-frontend/lib/contexts/InterviewContext';
 import { withPreferencesHOC } from 'evolution-frontend/lib/components/hoc/WithPreferencesHoc';
